requirements.txt
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
future==0.18.2
huepy==1.2.1
idna==2.10
imageio==2.8.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
instabot==0.117.0
mock==4.0.2
more-itertools==8.4.0
moviepy==1.0.3
numpy==1.19.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
proglog==0.1.9
py==1.9.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==5.4.3
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
responses==0.10.15
schedule==0.6.0
six==1.15.0
tqdm==4.47.0
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
